+++ lib/features/accounts/presentation/controllers/account_list_notifier.g.dart
@@ -383,4 +383,4 @@ final class AccountMapProvider
   }
 }
 
-String _$accountMapHash() => r'8242408973b46bcbd3d0efdc2c9ca94b051efae5';
+String _$accountMapHash() => r'32c36b0eeff5a0a4af45e9f24bd4d85c4abb6c91';