+++ lib/features/backup/presentation/sheets/backup_password_sheet.dart
@@ -14,6 +14,7 @@ Future<String?> showBackupPasswordSheet(
 }) async {
   return await showPokaSheet<String>(
     context: context,
+    persistent: true,
     builder: (context) =>
         _BackupPasswordSheet(isBackup: isBackup, onValidateRestore: onValidateRestore, filePath: filePath),
   );