+++ lib/features/categories/presentation/widgets/forms/category_form_sheet.dart
@@ -29,6 +29,7 @@ class CategoryFormSheet extends HookConsumerWidget {
   }) {
     return showPokaSheet(
       context: context,
+      persistent: true,
       builder: (context) => CategoryFormSheet(
         initialCategory: category,
         parentId: parentId,